creds (not finished yet)

+++ app/bot_telegram.py
@@ -5,7 +5,13 @@ from telegram.ext import Updater
 from telegram import Update
 from telegram.ext import CommandHandler, CallbackQueryHandler, MessageHandler, Filters, CallbackContext
 from telegram import InlineKeyboardButton, InlineKeyboardMarkup
+from collections import defaultdict
+import json
 import logging
+
+
+BOT_TOKEN = '1579751582:AAEcot5v5NLyxXB1uFYQiBCyvBAsKOzGGsU'
+CHAIN_LIST = ['waitrose', 'tesco', 'sainsbury']
 
 
 class BotState:
@@ -17,6 +23,17 @@ class BotState:
     is_waiting_password = False
     is_waiting_cvv = False
 
+    @staticmethod
+    def get_creds(chat_id):
+        with open('creds.json') as json_file:
+            data = json.load(json_file)
+        return data.get(str(chat_id), None)
+
+    @staticmethod
+    def update_creds(data):
+        with open('creds.json', 'w') as outfile:
+            json.dump(data, outfile)
+
     @classmethod
     def change_state(cls, display_name):
         if display_name in cls.state_list:
@@ -25,7 +42,10 @@ class BotState:
         else:
             cls.state_list.append(display_name)
 
-        logging.info(cls.state_list)
+        logging.debug(cls.state_list)
+
+
+bot_state = defaultdict(BotState)
 
 
 class Menu:
@@ -42,17 +62,19 @@ class Menu:
             c.parent = self
             c.register(dispatcher)
 
-    def display(self, bot, update):
+    def display(self, update, callback_context):
         logging.debug(f'self.display_name {self.display_name}')
         logging.debug(f'self.keyboard() {self._keyboard()}')
 
-        BotState.change_state(self.display_name)
-
-        bot.callback_query.message.edit_text(self.display_name, reply_markup=self._keyboard())
+        bs = bot_state[update.callback_query.message.chat.id]
+        bs.change_state(self.display_name)
+        update.callback_query.message.edit_text(self.display_name, reply_markup=self._keyboard())
 
     def create(self, bot, update, message=None):
         if not message:
             message = bot.message
+        bs = bot_state[message.chat.id]
+        bs.change_state(self.display_name)
         message.reply_text(self.display_name, reply_markup=self._keyboard())
 
     def _keyboard(self):
@@ -63,11 +85,23 @@ class Menu:
 
 
 class LoginMenu(Menu):
+    def __init__(self, display_name: str, children: list, filter_menu: Menu):
+        super().__init__(display_name, children)
+        self.filter_menu = filter_menu
+
     def display(self, bot, update):
-        BotState.change_state(self.display_name)
-        BotState.last_message = bot.callback_query.message.reply_text('Please enter your login')
-        bot.callback_query.message.delete()
-        BotState.is_waiting_login = True
+        bs = bot_state[bot.callback_query.message.chat.id]
+
+        creds = bs.get_creds(bot.callback_query.message.chat.id)
+
+        bs.change_state(self.display_name)
+
+        if not creds:
+            bs.last_message = bot.callback_query.message.reply_text('Please enter your login')
+            bot.callback_query.message.delete()
+            bs.is_waiting_login = True
+        else:
+            self.filter_menu.display(bot, update)
 
 
 class Bot:
@@ -96,8 +130,8 @@ class Bot:
 
             self.m_filter[chain] = Menu('Filters', [m_filter_slots, m_show_all_available_slots])
 
-            m_book_slot = LoginMenu('Book slot', [])
-            m_book_slot_and_checkout = LoginMenu('Book slot and checkout', [])
+            m_book_slot = LoginMenu('Book slot', [], self.m_filter[chain])
+            m_book_slot_and_checkout = LoginMenu('Book slot and checkout', [], self.m_filter[chain])
 
             m_chain = Menu(chain.capitalize(), [m_book_slot, m_book_slot_and_checkout])
             self.m_filter[chain].parent = m_chain
@@ -114,28 +148,39 @@ class Bot:
         self.updater.start_polling()
 
     def handle_text(self, update: Update, context: CallbackContext):
+        # if update.message.text.lower() == 'чмок':
+        #     self.updater.bot.send_sticker(update.message.chat.id, 'CAADAgADZgkAAnlc4gmfCor5YbYYRAI')
+
+        bs = bot_state[update.message.chat.id]
+
+        creds = bs.get_creds(update.message.chat.id)
+
         # an invitation
-        if BotState.last_message:
-            BotState.last_message.delete()
+        if bs.last_message:
+            bs.last_message.delete()
 
-        if BotState.is_waiting_login:
-            BotState.last_message = update.message.reply_text('Please, enter password')
-            BotState.is_waiting_login = False
-            BotState.is_waiting_password = True
-        elif BotState.is_waiting_password:
-            BotState.last_message = update.message.reply_text('Please, enter cvv')
-            BotState.is_waiting_password = False
-            BotState.is_waiting_cvv = True
-        elif BotState.is_waiting_cvv:
-            BotState.is_waiting_cvv = False
-            self.m_filter[BotState.state_list[0].lower()].create(None, None, update.message)
-            BotState.change_state('Filters')
+        if not creds:
+            if bs.is_waiting_login:
+                bs.last_message = update.message.reply_text('Please, enter password')
+                bs.is_waiting_login = False
+                bs.is_waiting_password = True
+            elif bs.is_waiting_password:
+                bs.last_message = update.message.reply_text('Please, enter cvv')
+                bs.is_waiting_password = False
+                bs.is_waiting_cvv = True
+            elif bs.is_waiting_cvv:
+                bs.is_waiting_cvv = False
+                self.m_filter[bs.state_list[1].lower()].create(None, None, update.message)
+                bs.change_state('Filters')
 
-        # user's input
-        update.message.delete()
+            # user's input
+            update.message.delete()
+        else:
+            self.m_filter[bs.state_list[1].lower()].create(None, None, update.message)
+            bs.change_state('Filters')
 
 
 if __name__ == '__main__':
     logging.basicConfig(format='%(levelname)s:%(message)s', level=logging.INFO)
-    b = Bot('1579751582:AAEcot5v5NLyxXB1uFYQiBCyvBAsKOzGGsU', ['waitrose', 'tesco'])
+    b = Bot(BOT_TOKEN, CHAIN_LIST)
     b.run()
